+++ rust/src/parser/src/unified/hy/hy_v4.rs
@@ -3,35 +3,34 @@
 
 use vllm_tokenizer::DynTokenizer;
 
-use super::{
-    CombinedParser, Result, UnifiedParser, UnifiedParserOutput, detect_hy_token_suffix, token_id,
-};
+use super::detect_hy_token_suffix;
 use crate::reasoning::HyReasoningParser;
 use crate::tool::{HyDialect, HyToolMarkers, HyToolParser, StructuralTagBuilder, Tool};
+use crate::unified::{CombinedParser, Result, UnifiedParser, UnifiedParserOutput, token_id};
 
-/// Unified reasoning and tool parser for HY3 output.
-pub struct HyV3UnifiedParser {
+/// Unified reasoning and tool parser for HY4 output.
+pub struct HyV4UnifiedParser {
     inner: CombinedParser,
 }
 
-impl HyV3UnifiedParser {
-    /// Create a HY3 parser using the suffix encoded in tokenizer added tokens.
+impl HyV4UnifiedParser {
+    /// Create a HY4 parser using the suffix encoded in tokenizer added tokens.
     pub fn new(tools: &[Tool], tokenizer: DynTokenizer) -> Result<Self> {
         let suffix = detect_hy_token_suffix(tokenizer.as_ref());
-        let markers = HyToolMarkers::new(&suffix, HyDialect::V3);
+        let markers = HyToolMarkers::new(&suffix, HyDialect::V4);
         for marker in markers.iter() {
             token_id(tokenizer.as_ref(), marker)?;
         }
 
         let reasoning = HyReasoningParser::new(tokenizer, &suffix)?;
-        let tool = HyToolParser::new(tools, &suffix, HyDialect::V3);
+        let tool = HyToolParser::new(tools, &suffix, HyDialect::V4);
         Ok(Self {
             inner: CombinedParser::new(Some(Box::new(reasoning)), Some(Box::new(tool))),
         })
     }
 }
 
-impl UnifiedParser for HyV3UnifiedParser {
+impl UnifiedParser for HyV4UnifiedParser {
     fn create(tools: &[Tool], tokenizer: DynTokenizer) -> Result<Box<dyn UnifiedParser>>
     where
         Self: Sized + 'static,
@@ -79,7 +78,7 @@ mod tests {
         FunctionDefinition, FunctionToolParam, ToolChoice, ToolParam, build_structural_tag,
     };
 
-    use super::{HyV3UnifiedParser, UnifiedParser};
+    use super::{HyV4UnifiedParser, UnifiedParser};
     use crate::tool::Tool;
     use crate::unified::{UnifiedParserEvent, UnifiedParserOutput};
 
@@ -91,7 +90,6 @@ mod tests {
             "</tool_calls:opensource>",
             "<tool_call:opensource>",
             "</tool_call:opensource>",
-            "<tool_sep:opensource>",
             "<arg_key:opensource>",
             "</arg_key:opensource>",
             "<arg_value:opensource>",
@@ -105,29 +103,38 @@ mod tests {
     }
 
     fn tools() -> Vec<Tool> {
-        vec![Tool {
-            name: "get_weather".to_string(),
-            description: None,
-            parameters: json!({
-                "type": "object",
-                "properties": { "city": { "type": "string" } },
-            }),
-            strict: None,
-        }]
+        vec![
+            Tool {
+                name: "get_weather".to_string(),
+                description: None,
+                parameters: json!({
+                    "type": "object",
+                    "properties": { "city": { "type": "string" } },
+                }),
+                strict: None,
+            },
+            Tool {
+                name: "get_current_date".to_string(),
+                description: None,
+                parameters: json!({ "type": "object", "properties": {} }),
+                strict: None,
+            },
+        ]
     }
 
     #[test]
-    fn parses_suffixed_reasoning_and_tool_call_as_one_stream() {
+    fn parses_suffixed_reasoning_and_compact_tool_calls_as_one_stream() {
         let tokenizer = Arc::new(tokenizer());
         let think_start_id = tokenizer.token_to_id("<think:opensource>").unwrap();
-        let mut parser = HyV3UnifiedParser::new(&tools(), tokenizer).unwrap();
+        let mut parser = HyV4UnifiedParser::new(&tools(), tokenizer).unwrap();
         parser.initialize(&[think_start_id]).unwrap();
 
         let chunks = [
             "reasoning</think:open",
             "source>answer<tool_calls:opensource><tool_call:opensource>get_weather",
-            "<tool_sep:opensource><arg_key:opensource>city</arg_key:opensource>",
+            "<arg_key:opensource>city</arg_key:opensource>",
             "<arg_value:opensource>Beijing</arg_value:opensource>",
+            "</tool_call:opensource><tool_call:opensource>get_current_date",
             "</tool_call:opensource></tool_calls:opensource>",
         ];
         let mut output = UnifiedParserOutput::default();
@@ -146,13 +153,18 @@ mod tests {
                     name: Some("get_weather".to_string()),
                     arguments: r#"{"city":"Beijing"}"#.to_string(),
                 }),
+                UnifiedParserEvent::ToolCall(crate::tool::ToolCallDelta {
+                    tool_index: 1,
+                    name: Some("get_current_date".to_string()),
+                    arguments: "{}".to_string(),
+                }),
             ]
         );
     }
 
     #[test]
-    fn structural_tag_uses_tokenizer_detected_suffix() {
-        let parser = HyV3UnifiedParser::new(&tools(), Arc::new(tokenizer())).unwrap();
+    fn structural_tag_uses_compact_hy4_skeleton() {
+        let parser = HyV4UnifiedParser::new(&tools(), Arc::new(tokenizer())).unwrap();
         let structural_tools = [ToolParam::Function(FunctionToolParam::new(
             FunctionDefinition::new("get_weather").with_parameters(json!({
                 "type": "object",
@@ -172,7 +184,11 @@ mod tests {
         .unwrap();
 
         assert!(tag.contains("<tool_calls:opensource>"));
-        assert!(tag.contains("<arg_value:opensource>"));
-        assert!(!tag.contains("glm_xml"));
+        assert!(tag.contains("<tool_call:opensource>get_weather"));
+        assert!(tag.contains("<arg_key:opensource>"));
+        assert!(tag.contains("</tool_call:opensource>"));
+        assert!(tag.contains("</tool_calls:opensource>"));
+        assert!(!tag.contains("tool_sep"));
+        assert!(!tag.contains(r#"\n"#));
     }
 }
