+++ rust/src/parser/src/unified/mod.rs
@@ -5,15 +5,13 @@
 
 mod combined;
 mod gemma4;
-mod hy_v3;
-mod hy_v4;
+mod hy;
 mod inkling;
 mod kimi_k3;
 
 pub use combined::CombinedParser;
 pub use gemma4::Gemma4UnifiedParser;
-pub use hy_v3::HyV3UnifiedParser;
-pub use hy_v4::HyV4UnifiedParser;
+pub use hy::{HyV3UnifiedParser, HyV4UnifiedParser};
 pub use inkling::InklingUnifiedParser;
 pub use kimi_k3::{KimiK3StructuralTagBuilder, KimiK3UnifiedParser};
 use thiserror::Error;
@@ -24,37 +22,6 @@ use crate::reasoning::ReasoningError;
 use crate::tool::{
     StructuralTagBuilder, Tool, ToolCallDelta, ToolParserError, ToolParserEvent, ToolParserOutput,
 };
-
-const HY_MARKER_STEMS: &[&str] = &[
-    "think",
-    "tool_calls",
-    "tool_call",
-    "tool_sep",
-    "arg_key",
-    "arg_value",
-];
-
-/// Detect the HY structural-token suffix from tokenizer added vocabulary.
-fn detect_hy_token_suffix(tokenizer: &dyn vllm_tokenizer::Tokenizer) -> String {
-    tokenizer
-        .added_vocab()
-        .iter()
-        .filter_map(|(token, id)| hy_marker_suffix(token).map(|suffix| (*id, suffix)))
-        .min_by_key(|(id, _)| *id)
-        .map(|(_, suffix)| suffix.to_string())
-        .unwrap_or_default()
-}
-
-/// Extract the suffix from one opening or closing HY structural token.
-fn hy_marker_suffix(token: &str) -> Option<&str> {
-    let body = token.strip_prefix('<')?.strip_suffix('>')?;
-    let body = body.strip_prefix('/').unwrap_or(body);
-
-    HY_MARKER_STEMS.iter().find_map(|stem| {
-        let suffix = body.strip_prefix(stem)?;
-        (suffix.is_empty() || suffix.starts_with(':')).then_some(suffix)
-    })
-}
 
 /// Result alias for unified parser operations.
 pub type Result<T> = std::result::Result<T, UnifiedParserError>;
